requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aioitertools==0.12.0
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
arize-phoenix==11.30.0
arize-phoenix-client==1.16.0
arize-phoenix-evals==0.29.0
arize-phoenix-otel==0.13.0
art==6.5
arxiv==2.2.0
asttokens==3.0.0
attrs==25.3.0
authlib==1.6.3
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
chunkr-ai==0.3.7
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cryptography==45.0.7
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
email-validator==2.3.0
et-xmlfile==2.0.0
executing==2.2.1

fastapi==0.116.1
feedparser==6.0.11
fonttools==4.59.2
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.32.0
googleapis-common-protos==1.70.0
graphql-core==3.2.6
greenlet==3.2.4
griffe==1.13.0
grpc-interceptor==0.15.4
grpcio==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hydra-core==1.3.2
idna==3.10
importlib-metadata==8.7.0
ipython==9.5.0
ipython-pygments-lexers==1.1.1
jedi==0.19.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpath-ng==1.7.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.1.7
mcp==1.13.1
mdurl==0.1.2
multidict==6.6.4
neo4j==5.28.2
nest-asyncio==1.6.0
numpy==2.3.2
ollama==0.5.3
omegaconf==2.3.0
openai==1.102.0
openai-agents==0.2.10
openinference-instrumentation==0.1.37
openinference-instrumentation-openai==0.1.31
openinference-instrumentation-openai-agents==1.2.0
openinference-semantic-conventions==0.1.21
openpyxl==3.1.5
opentelemetry-api==1.36.0
opentelemetry-exporter-otlp==1.36.0
opentelemetry-exporter-otlp-proto-common==1.36.0
opentelemetry-exporter-otlp-proto-grpc==1.36.0
opentelemetry-exporter-otlp-proto-http==1.36.0
opentelemetry-instrumentation==0.57b0
opentelemetry-proto==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
ply==3.11
prometheus-client==0.22.1
prompt-toolkit==3.0.52
propcache==0.3.2
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pygments==2.19.2
pyparsing==3.2.3
pystache==0.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pyyaml==6.0.2
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
rich==14.1.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.1
scipy==1.16.1
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.43
sqlean-py==3.49.1
sqlmodel==0.0.24
sse-starlette==3.0.2
stack-data==0.6.3
starlette==0.47.3
strawberry-graphql==0.270.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
toml==0.10.2
tqdm==4.67.1
traitlets==5.14.3
types-requests==2.32.4.20250809
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
wikipedia-api==0.8.1
wrapt==1.17.3
yarl==1.20.1
utcp>=1.0.0
utcp-http>=1.0.0
utcp-mcp>=1.0.0

zipp==3.23.0
pytest==8.3.5
